ai_research_lab/requirements.txt
crewai
crewai-tools
python-dotenv
fastapi
uvicorn[standard]
langchain-google-genai
python-multipart
jinja2
guardrails-ai
langchain
langsmith
